+++ sdk/src/test/java/com/hedera/hashgraph/sdk/AccountIdTest.java
@@ -111,6 +111,13 @@ class AccountIdTest {
     }
 
     @Test
+    void malformedAliasKey3() {
+        assertThrows(IllegalArgumentException.class, () -> {
+            AccountId.fromString("0.0.303a300506032b6570032100114e6abc371b82dab5c15ea149f02d34a012087b163516dd70f44acafabf7777");
+        });
+    }
+
+    @Test
     void fromStringWithAliasKey() {
         SnapshotMatcher.expect(AccountId.fromString("0.0.302a300506032b6570032100114e6abc371b82dab5c15ea149f02d34a012087b163516dd70f44acafabf7777")).toMatchSnapshot();
     }