+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+ * `AccountId.aliasKey`, including `AccountId.[to|from]String()` support.
+ * `PublicKey.toAccountId()` and `PrivateKey.toAccountId()`.
+ * `aliasKey` fields in `TransactionRecord` and `AccountInfo`.
+ * New alias key response codes.
+
 ## v2.5.0
 
 ### Added